+++ src/js/components/dashboard/graph/BarChartXAxis.jsx
@@ -19,10 +19,33 @@ const propTypes = {
 
 export default class BarChartXAxis extends React.Component {
     render() {
-        const labels = this.props.items.map((item) => (
+        const timePeriodLabels = [];
+        this.props.items.forEach((yearItem) => {
+            yearItem.timePeriodItems.forEach((periodItem) => (
+                timePeriodLabels.push(
+                    <BarChartXAxisItem
+                        key={`${yearItem.value} ${periodItem.value}`}
+                        label={periodItem.label}
+                        x={periodItem.x}
+                        y={periodItem.y} />
+                )
+            ));
+        });
+        const yearLines = this.props.items.map((item) => (
+            <line
+                key={item.value}
+                className="x-axis"
+                x1={item.startX}
+                x2={item.endX}
+                y1={40}
+                y2={40} />
+        ));
+        const yearLabels = this.props.items.map((yearItem) => (
             <BarChartXAxisItem
-                {...item}
-                key={item.value} />
+                key={yearItem.value}
+                label={`FY ${yearItem.label}`}
+                x={yearItem.x}
+                y={yearItem.y} />
         ));
 
         return (
@@ -40,7 +63,15 @@ export default class BarChartXAxis extends React.Component {
                 <g
                     className="axis-labels"
                     transform={`translate(${this.props.labelGroup.x},${this.props.labelGroup.y})`}>
-                    {labels}
+                    {timePeriodLabels}
+                </g>
+                <g transform={`translate(${this.props.lineGroup.x}, ${this.props.lineGroup.y})`}>
+                    {yearLines}
+                </g>
+                <g
+                    className="axis-labels"
+                    transform={`translate(${this.props.labelGroup.x},${this.props.labelGroup.y})`}>
+                    {yearLabels}
                 </g>
             </g>
         );