+++ tests/redux/dashboard/dashboardFiltersReducer-test.js
@@ -12,22 +12,22 @@ describe('dashboardFiltersReducer', () => {
         const action = {
             type: 'UPDATE_FILTER_SET',
             dashboardType: 'historical',
-            filterType: 'quarters',
-            filterValue: 2
+            filterType: 'fy',
+            filterValue: 2018
         };
 
         it('should add the value if it does not currently exist in the filter', () => {
             const updatedState = dashboardFiltersReducer(undefined, action);
 
-            expect(updatedState.historical.quarters).toEqual(new Set([2]));
+            expect(updatedState.historical.fy).toEqual(new Set([2018]));
         });
 
         it('should remove the value if already exists in the filter', () => {
             const startingState = cloneDeep(initialState);
-            startingState.historical.quarters = new Set([2]);
+            startingState.historical.fy = new Set([2018]);
 
             const updatedState = dashboardFiltersReducer(startingState, action);
-            expect(updatedState.historical.quarters).toEqual(new Set());
+            expect(updatedState.historical.fy).toEqual(new Set());
         });
     });
     describe('UPDATE_GENERIC_FILTER', () => {
@@ -35,7 +35,10 @@ describe('dashboardFiltersReducer', () => {
             type: 'UPDATE_GENERIC_FILTER',
             dashboardType: 'historical',
             filterType: 'agency',
-            value: '123'
+            value: {
+                code: '123',
+                name: 'Test Agency'
+            }
         };
 
         it('should change the given filter to the provided value', () => {
@@ -43,22 +46,22 @@ describe('dashboardFiltersReducer', () => {
 
             state = dashboardFiltersReducer(state, action);
 
-            expect(state.historical.agency).toEqual('123');
+            expect(state.historical.agency).toEqual({ code: '123', name: 'Test Agency' });
         });
 
         it('should remove the value if it is the value already selected', () => {
             const startingState = cloneDeep(initialState);
-            startingState.historical.agency = '123';
+            startingState.historical.agency = { code: '123', name: 'Test Agency' };
 
             const updatedState = dashboardFiltersReducer(startingState, action);
-            expect(updatedState.historical.agency).toEqual('');
+            expect(updatedState.historical.agency).toEqual({ code: '', name: '' });
         });
     });
     describe('CLEAR_FILTER', () => {
         it('should reset the specified filter to its initial state', () => {
             const state = cloneDeep(initialState);
             state.historical = {
-                quarters: new Set([1, 3]),
+                period: 'Q1',
                 fy: new Set([2018, 2019]),
                 file: 'A',
                 rules: new Set(['X12', 'X34'])
@@ -75,14 +78,14 @@ describe('dashboardFiltersReducer', () => {
 
             expect(restoredState.historical.rules).toEqual(initialState.historical.rules);
             expect(restoredState.historical.file).toEqual('A');
-            expect(restoredState.historical.quarters).toEqual(new Set([1, 3]));
+            expect(restoredState.historical.period).toEqual('Q1');
         });
     });
     describe('CLEAR_DASHBOARD_FILTERS', () => {
         it('should reset the dashboard filters to their initial state', () => {
             const state = cloneDeep(initialState);
             state.historical = {
-                quarters: new Set([1, 3]),
+                period: 'Q2',
                 fy: new Set([2018, 2019]),
                 file: 'A',
                 rules: new Set(['X12', 'X34'])