+++ src/help/history.md
@@ -1,4 +1,12 @@
-#### June 22, 2021{section=changelog}
+#### July 13, 2021{section=changelog}
+In this release of the Broker, we:
+
+* Implemented DAIMS 2.1 guidance on submission of ObjectClass (OC) and DisasterEmergencyFundCode (DEFC) elements. 
+* DEFC ‘9’ is no longer valid for submission as a generic placeholder, only valid DEFC may be submitted.
+* Object Class format modified to match GTAS format.
+* Modified several validation rules impacted by DEFC and OC changes.
+
+#### June 22, 2021
 In this release of the Broker, we:
 
 * Updated rule A33 to properly account for FREC agencies and to match the logic of File A generation in associating TAS with Agency.